aipms-hackathon/requirements.txt
# AIPMS 36-Hour Hackathon - Requirements
# BIT Sindri Problem Statement PS-1: AI-Driven Predictive Maintenance

# === Core ML/Data Stack ===
pandas==2.2.0
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
torch==2.1.1
streamlit==1.32.0
plotly==5.18.0

# === IoT/Backend ===
paho-mqtt==1.6.1
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.23
APScheduler==3.10.4
requests==2.31.0

# === ML Explainability ===
shap==0.44.0

# === Data Processing ===
scipy==1.11.4

# === Configuration ===
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# === Testing ===
pytest==8.0.0
pytest-asyncio==0.23.0

# === Utilities ===
python-dateutil==2.8.2
